Split Dashboard.vue into components and remove dead PO UI
Fractal split
Dashboard.vue shrinks from 1,271 lines to ~395 by extracting each
section into a focused component under client/src/components/dashboard/:
- KpiSection.vue — 5 KPI cards with progress bars
- OrderHealthCard.vue — donut chart + metrics panel
- InventoryByCategoryCard.vue — horizontal bar chart
- InventoryShortagesCard.vue — backlog shortage table
- TopProductsCard.vue — top products table
Dashboard.vue keeps the orchestration role: data loading, filter
watching, modal state, and passing props/listening to events.

Dead purchase-order cleanup
The shortages card rendered Create PO / View PO buttons that emitted
events into state which was forwarded to a <PurchaseOrderModal> that
was never imported and whose .vue file does not exist — the buttons
set showPOModal but no modal ever observed it, so clicks were no-ops.
Removed the entire dead path:
- Dashboard.vue: <PurchaseOrderModal> tag, showPOModal/selectedBacklog
  ForPO/poModalMode refs, openPOModal/viewPO/handlePOCreated handlers
- InventoryShortagesCard.vue: Actions column, PO buttons, related
  emits, .po-button CSS
- Locales: dashboard.inventoryShortages.createPO, .viewPO, and the
  now-unused common.actions key
